discord_bot/requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
